Animatable

+++ Components/FavoritesComponent.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import Swipeout from 'react-native-swipeout';
+import * as Animatable from 'react-native-animatable';
 import { deleteFavorite } from '../redux/ActionCreators';
 import { Loading } from './LoadingComponent';
 import { baseUrl } from '../shared/baseUrl';
@@ -29,17 +30,36 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite campsite' + item.name + '?',
+                            [
+                                {
+                                    text: 'Cancel',
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: 'cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                    } 
                 }
             ];
             return(
                 <Swipeout right={rightButton} autoClose={true}>
-                    <ListItem
-                        title={item.name}
-                        subtitle={item.description}
-                        leftAvatar={{source: {uri: baseUrl + item.image}}}
-                        onPress={() => navigate('CampsiteInfo', {campsiteId: item.id})}
-                    />
+                    <Animatable.View animation='fadeInRightBig' duration={2000}>
+                        <ListItem
+                            title={item.name}
+                            subtitle={item.description}
+                            leftAvatar={{source: {uri: baseUrl + item.image}}}
+                            onPress={() => navigate('CampsiteInfo', {campsiteId: item.id})}
+                        />
+                    </Animatable.View>    
                 </Swipeout>    
             );
         };
